add test framework.

+++ final/include/myMemory.h
@@ -2,12 +2,11 @@
 #define _MY_MEMORY_H
 
 
-
-void *myAlloc(size_t);
+void *myAlloc(unsigned int);
 
 void myFree(void *);
 
-void *myReAlloc(void *, size_t);
+void *myReAlloc(void *, unsigned int);
